livekit-server/pyproject.toml
[project]
name = "livekit-token-server"
version = "0.1.0"
description = "Token server and static client for LiveKit collect-agent"
requires-python = ">=3.10"
dependencies = [
    "livekit-api>=1.1.0",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
